+++ example/cumulus-tf/cnm_response_task.tf
@@ -3,7 +3,7 @@ resource "null_resource" "get_cnmResponse" {
     always_run = timestamp()
   }
   provisioner "local-exec" {
-    command = "curl -s -L -o cnmResponse.zip https://github.com/podaac/cumulus-cnm-response-task/releases/download/v1.2.0/cnmResponse-1.2.0.zip"
+    command = "curl -s -L -o cnmResponse.zip https://github.com/podaac/cumulus-cnm-response-task/releases/download/v1.4.4/cnmResponse-1.4.4.zip"
   }
 }