+++ CHANGELOG.md
@@ -4,7 +4,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [v9.4.1] 2022-02-14 [BACKPORT]
+
+- **CUMULUS-2847**
+  - Update dynamo configuration to read from S3 instead of System Manager
+    Parameter Store
+  - Move api configuration initialization outside the lambda handler to
+    eliminate unneded S3 calls/require config on cold-start only
+  - Moved `ssh2` package from `@cumulus/common` to `@cumulus/sftp-client` and
+    upgraded package from `^0.8.7` to `^1.0.0` to address security vulnerability
+    issue in previous version.
+  - Fixed hyrax task package.json dev dependency
+  - Update CNM lambda dependencies for Core tasks
+    - cumulus-cnm-response-task: 1.4.4
+    - cumulus-cnm-to-granule: 1.5.4
+  - Whitelist ssh2 re: https://github.com/advisories/GHSA-652h-xwhf-q4h6
 
 ## [v9.4.0] 2021-08-13
 
@@ -16,7 +30,6 @@ required and involves a redirect to a different host (e.g.
 downloading files protected by Earthdata Login)
 
 ### Added
-
 - **CUMULUS-2591**
   - Adds `failedExecutionStepName` to failed execution's jsonb error records.
     This is the name of the Step Function step for the last failed event in the
@@ -4754,7 +4767,8 @@ Note: There was an issue publishing 1.12.0. Upgrade to 1.12.1.
 
 ## [v1.0.0] - 2018-02-23
 
-[unreleased]: https://github.com/nasa/cumulus/compare/v9.4.0...HEAD
+[unreleased]: https://github.com/nasa/cumulus/compare/v9.4.1...HEAD
+[v9.4.1]: https://github.com/nasa/cumulus/compare/v9.4.0...v9.4.1
 [v9.4.0]: https://github.com/nasa/cumulus/compare/v9.3.0...v9.4.0
 [v9.3.0]: https://github.com/nasa/cumulus/compare/v9.2.2...v9.3.0
 [v9.2.2]: https://github.com/nasa/cumulus/compare/v9.2.1...v9.2.2